// src/app/model/apa/performance.model.ts
export interface PerformanceModel {
    id: number | null;
    name: string | null;
    nameBn: string | null;
    code: string | null;
    programCode: string | null;
    poName: string | null;
    indexUnitCode: string | null;
    iuName: string | null;
    value: number | null;
    orderBy: number | null;
}